Update objects.py
STILL ERRORS WITH PERFORMANCE OF GA. TEMPORARY VERSION SUBJECT TO CHANGE
Changed genetic algorithm to not randomly create new creatures but mutate parents also if the fitnesses are low. Also implemented the stop for a dead population. There is still a problem as all stats are subject to changes when a pops ability to survive dramatically changes.